+++ src/requisition-view/requisition-view.controller.js
@@ -68,11 +68,20 @@
          * @type {String}
          *
          * @description
-         * Holds message key to display, depending on the requisition type (regular/emergency).
+         * Holds message key to display, depending on the requisition type (regular/emergency/report-only).
          */
-        vm.requisitionType = vm.requisition.emergency ?
-            'requisitionView.emergency' :
-            'requisitionView.regular';
+        vm.requisitionType = undefined;
+
+        /**
+         * @ngdoc property
+         * @propertyOf requisition-view.controller:RequisitionViewController
+         * @name requisitionTypeClass
+         * @type {String}
+         *
+         * @description
+         * Holds CSS class to use, depending on the requisition type (regular/emergency/report-only).
+         */
+        vm.requisitionTypeClass = undefined;
 
         /**
          * @ngdoc property
@@ -198,6 +207,7 @@
          * Initialization method of the RequisitionViewController.
          */
         function onInit() {
+            setTypeAndClass();
             vm.displaySubmitButton = canSubmit && !vm.requisition.program.skipAuthorization;
             vm.displaySubmitAndAuthorizeButton = canSubmit && vm.requisition.program.skipAuthorization;
             vm.displayAuthorizeButton = canAuthorize;
@@ -205,6 +215,19 @@
             vm.displayApproveAndRejectButtons = canApproveAndReject;
             vm.displaySkipButton = canSkip;
             vm.displaySyncButton = canSync;
+        }
+
+        function setTypeAndClass() {
+            if (vm.requisition.emergency) {
+                vm.requisitionType = 'requisitionView.emergency';
+                vm.requisitionTypeClass = 'emergency';
+            } else if (vm.requisition.reportOnly) {
+                vm.requisitionType = 'requisitionView.reportOnly';
+                vm.requisitionTypeClass = 'report-only';
+            } else {
+                vm.requisitionType = 'requisitionView.regular';
+                vm.requisitionTypeClass = 'regular';
+            }
         }
 
         /**